+++ controller_2023/controller/data/LED.json
@@ -1,9 +1,24 @@
 {
   "led1": [
     {
-      "start": 400,
+      "start": 150,
+      "fade": true,
+      "status": [
+          [40, 0, 0, 10],
+          [40, 0, 0, 10],
+          [40, 0, 0, 10],
+          [40, 0, 0, 10]
+      ]
+    },
+    {
+      "start": 300,
       "fade": false,
-      "status": [[10, 20, 40, 10]]
+      "status": [
+          [0, 40, 0, 10],
+          [0, 40, 0, 10],
+          [0, 40, 0, 10],
+          [0, 40, 0, 10]
+      ]
     }
   ],
   "led2": [
@@ -11,82 +26,16 @@
       "start": 0,
       "fade": true,
       "status": [
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10]
+        [0, 0, 40, 10],
+        [0, 0, 40, 10],
+        [0, 0, 40, 10],
+        [0, 0, 40, 10]
       ]
     },
     {
-      "start": 1792,
+      "start": 300,
       "fade": false,
       "status": [
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
         [0, 0, 0, 10],
         [0, 0, 0, 10],
         [0, 0, 0, 10],
@@ -97,84 +46,38 @@
   "led3": [
     {
       "start": 0,
-      "fade": false,
+      "fade": true,
       "status": [
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 20, 40, 10],
-        [10, 120, 40, 10],
-        [101, 20, 40, 10]
+        [40, 40, 40, 10],
+        [40, 40, 40, 10],
+        [40, 40, 40, 10],
+        [40, 40, 40, 10]
       ]
     },
     {
-      "start": 1792,
+      "start": 60,
       "fade": false,
       "status": [
         [0, 0, 0, 10],
         [0, 0, 0, 10],
         [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
-        [0, 0, 0, 10],
+        [0, 0, 0, 10]
+      ]
+    },
+    {
+      "start": 120,
+      "fade": false,
+      "status": [
+        [40, 40, 40, 10],
+        [40, 40, 40, 10],
+        [40, 40, 40, 10],
+        [40, 40, 40, 10]
+      ]
+    },
+    {
+      "start": 180,
+      "fade": false,
+      "status": [
         [0, 0, 0, 10],
         [0, 0, 0, 10],
         [0, 0, 0, 10],
