+++ controller_2023/controller/data/OF.json
@@ -4,96 +4,52 @@
     "start": 0,
     "fade": true,
     "status": {
-      "of1": [255, 0, 0, 10],
-      "of2": [30, 0, 0, 10],
-      "of3": [30, 0, 0, 10],
-      "of4": [30, 0, 0, 10],
-      "of5": [30, 0, 0, 10],
-      "of6": [30, 0, 0, 10],
-      "of7": [30, 0, 0, 10],
-      "of8": [30, 0, 0, 10],
-      "of9": [30, 0, 0, 10],
-      "of10": [30, 0, 0, 10],
-      "of11": [30, 0, 0, 10],
-      "of12": [30, 0, 0, 10],
-      "of13": [30, 0, 0, 10],
-      "of14": [30, 0, 0, 10],
-      "of15": [30, 0, 0, 10],
-      "of16": [30, 0, 0, 10],
-      "of17": [30, 0, 0, 10],
-      "of18": [30, 0, 0, 10],
-      "of19": [30, 0, 0, 10],
-      "of20": [30, 0, 0, 10],
-      "of21": [30, 0, 0, 10],
-      "of22": [30, 0, 0, 10],
-      "of23": [30, 0, 0, 10],
-      "of24": [30, 0, 0, 10],
-      "of25": [30, 0, 0, 10],
-      "of26": [30, 0, 0, 10]
+      "of1": [255, 255, 255, 10],
+      "of2": [255, 255, 255, 10],
+      "of3": [255, 255, 255, 10],
+      "of4": [255, 255, 255, 10],
+      "of5": [255, 255, 255, 10],
+      "of6": [255, 255, 255, 10],
+      "of7": [255, 255, 255, 10],
+      "of8": [255, 255, 255, 10],
+      "of9": [255, 255, 255, 10],
+      "of10": [255, 255, 255, 10],
+      "of11": [255, 255, 255, 10],
+      "of12": [255, 255, 255, 10],
+      "of13": [255, 255, 255, 10],
+      "of14": [255, 255, 255, 10],
+      "of15": [255, 255, 255, 10],
+      "of16": [255, 255, 255, 10],
+      "of17": [255, 255, 255, 10],
+      "of18": [255, 255, 255, 10],
+      "of19": [255, 255, 255, 10],
+      "of20": [255, 255, 255, 10]
     }
   },
   {
     "start": 10000,
     "fade": true,
     "status": {
-      "of1": [0, 255, 0, 0],
-      "of2": [0, 30, 0, 0],
-      "of3": [0, 30, 0, 0],
-      "of4": [0, 30, 0, 0],
-      "of5": [0, 30, 0, 0],
-      "of6": [0, 30, 0, 0],
-      "of7": [0, 30, 0, 0],
-      "of8": [0, 30, 0, 0],
-      "of9": [0, 30, 0, 0],
-      "of10": [0, 30, 0, 0],
-      "of11": [0, 30, 0, 0],
-      "of12": [0, 30, 0, 0],
-      "of13": [0, 30, 0, 0],
-      "of14": [0, 30, 0, 0],
-      "of15": [0, 30, 0, 0],
-      "of16": [0, 30, 0, 0],
-      "of17": [0, 30, 0, 0],
-      "of18": [0, 30, 0, 0],
-      "of19": [0, 30, 0, 0],
-      "of20": [0, 30, 0, 0],
-      "of21": [0, 30, 0, 0],
-      "of22": [0, 30, 0, 0],
-      "of23": [0, 30, 0, 0],
-      "of24": [0, 30, 0, 0],
-      "of25": [0, 30, 0, 0],
-      "of26": [0, 30, 0, 0]
-    }
-  },
-  {
-    "start": 12000,
-    "fade": false,
-    "status": {
-      "of1": [0, 0, 0, 0],
-      "of2": [0, 0, 0, 0],
-      "of3": [0, 0, 0, 0],
-      "of4": [0, 0, 0, 0],
-      "of5": [0, 0, 0, 0],
-      "of6": [0, 0, 0, 0],
-      "of7": [0, 0, 0, 0],
-      "of8": [0, 0, 0, 0],
-      "of9": [0, 0, 0, 0],
-      "of10": [0, 0, 0, 0],
-      "of11": [0, 0, 0, 0],
-      "of12": [0, 0, 0, 0],
-      "of13": [0, 0, 0, 0],
-      "of14": [0, 0, 0, 0],
-      "of15": [0, 0, 0, 0],
-      "of16": [0, 0, 0, 0],
-      "of17": [0, 0, 0, 0],
-      "of18": [0, 0, 0, 0],
-      "of19": [0, 0, 0, 0],
-      "of20": [0, 0, 0, 0],
-      "of21": [0, 0, 0, 0],
-      "of22": [0, 0, 0, 0],
-      "of23": [0, 0, 0, 0],
-      "of24": [0, 0, 0, 0],
-      "of25": [0, 0, 0, 0],
-      "of26": [0, 0, 0, 0]
+      "of1": [255, 255, 255, 0],
+      "of2": [255, 255, 255, 0],
+      "of3": [255, 255, 255, 0],
+      "of4": [255, 255, 255, 0],
+      "of5": [255, 255, 255, 0],
+      "of6": [255, 255, 255, 0],
+      "of7": [255, 255, 255, 0],
+      "of8": [255, 255, 255, 0],
+      "of9": [255, 255, 255, 0],
+      "of10": [255, 255, 255, 0],
+      "of11": [255, 255, 255, 0],
+      "of12": [255, 255, 255, 0],
+      "of13": [255, 255, 255, 0],
+      "of14": [255, 255, 255, 0],
+      "of15": [255, 255, 255, 0],
+      "of16": [255, 255, 255, 0],
+      "of17": [255, 255, 255, 0],
+      "of18": [255, 255, 255, 0],
+      "of19": [255, 255, 255, 0],
+      "of20": [255, 255, 255, 0]
     }
   }
 ]
