+++ controller_2023/controller/data/testName.json
@@ -20,26 +20,20 @@
     "of17": 4,
     "of18": 11,
     "of19": 18,
-    "of20": 12,
-    "of21": 8,
-    "of22": 10,
-    "of23": 15,
-    "of24": 16,
-    "of25": 22,
-    "of26": 24
+    "of20": 12
   },
   "LEDPARTS": {
     "led1": {
       "id": 0,
-      "len": 1
+      "len": 4
     },
     "led2": {
       "id": 1,
-      "len": 37
+      "len": 4
     },
     "led3": {
       "id": 2,
-      "len": 37
+      "len": 4
     }
   }
 }
